新文件：   .vscode/c_cpp_properties.json 	修改：     .vscode/settings.json 	修改：     Makefile 	修改：     README.md 	修改：     apps/Makefile 	新文件：   apps/apilib/apilib.obj 	新文件：   apps/apilib/libapi.a 	新文件：   apps/include/ctype.h 	新文件：   apps/include/errno.h 	删除：     apps/include/napi.h 	新文件：   apps/include/napi/consio/cons.h 	新文件：   apps/include/napi/consio/consi.h 	新文件：   apps/include/napi/consio/conso.h 	新文件：   apps/include/napi/file/file.h 	新文件：   apps/include/napi/memory.h 	新文件：   apps/include/napi/process.h 	新文件：   apps/include/napi/time.h 	新文件：   apps/include/stdbool.h 	修改：     apps/include/stddef.h 	修改：     apps/include/stdint.h 	修改：     apps/include/stdio.h 	修改：     apps/include/stdlib.h 	修改：     apps/include/string.h 	新文件：   apps/include/time.h 	删除：     apps/include/unistd.h 	新文件：   apps/klotski/Makefile 	新文件：   apps/klotski/klotski.bin 	新文件：   apps/klotski/klotski.c 	新文件：   apps/klotski/klotski.h 	新文件：   apps/klotski/klotski.o 	新文件：   apps/klotski/main.c 	新文件：   apps/klotski/main.o 	新文件：   apps/klotski/rand.c 	新文件：   apps/klotski/rand.o 	修改：     apps/lib/Makefile 	新文件：   apps/lib/ctype.c 	修改：     apps/lib/malloc.c 	新文件：   apps/lib/malloc.obj 	新文件：   apps/lib/rand.c 	删除：     apps/lib/sprintf.c 	修改：     apps/lib/stdio.c 	新文件：   apps/lib/stdlib.c 	修改：     apps/lib/string.c 	新文件：   apps/lib/time.c 	修改：     apps/myapp/Makefile 	修改：     apps/myapp/myapp.c 	删除：     apps/myapp2/Makefile 	删除：     apps/myapp2/myapp2.bin 	删除：     apps/myapp2/myapp2.c 	删除：     apps/myapp2/myapp2.elf 	删除：     apps/myapp2/myapp2.obj 	修改：     apps/napilib/Makefile 	删除：     apps/napilib/napi000.asm 	删除：     apps/napilib/napi001.asm 	删除：     apps/napilib/napi002.asm 	删除：     apps/napilib/napi003.asm 	删除：     apps/napilib/napi004.asm 	删除：     apps/napilib/napi005.asm 	删除：     apps/napilib/napi006.asm 	删除：     apps/napilib/napi007.asm 	删除：     apps/napilib/napi008.asm 	删除：     apps/napilib/napi009.asm 	删除：     apps/napilib/napi010.asm 	删除：     apps/napilib/napi011.asm 	新文件：   apps/napilib/napilib.asm 	删除：     apps/newyear/Makefile 	删除：     apps/newyear/newyear.asm 	删除：     apps/newyear/newyear.bin 	修改：     apps/nsh/nsh.bin 	修改：     apps/start/start.c 	修改：     config.txt 	修改：     include/load.inc 	修改：     kernel/ELF/elf.c 	修改：     kernel/Makefile 	修改：     kernel/disk/Makefile 	删除：     kernel/disk/disk.c 	修改：     kernel/disk/vdisk.c 	修改：     kernel/drivers/Makefile 	修改：     kernel/drivers/acpi/Makefile 	修改：     kernel/drivers/acpi/acpi.c 	新文件：   kernel/drivers/acpi/hpet.c 	新文件：   kernel/drivers/audio/Makefile 	新文件：   kernel/drivers/audio/audio.c 	新文件：   kernel/drivers/audio/sb16/sb16.c 	修改：     kernel/drivers/com/com.c 	修改：     kernel/drivers/cpu/cpu.c 	修改：     kernel/drivers/dbuffer/dbuffer.c 	修改：     kernel/drivers/dma/dma.c 	修改：     kernel/drivers/fdc/fdc.c 	删除：     kernel/drivers/hd/Makefile 	删除：     kernel/drivers/hd/hd.c 	修改：     kernel/drivers/ide/ide.c 	新文件：   kernel/drivers/interrupt/Makefile 	新文件：   kernel/drivers/interrupt/apic.c 	修改：     kernel/drivers/io/Makefile 	删除：     kernel/drivers/io/io.asm 	新文件：   kernel/drivers/io/io.c 	修改：     kernel/drivers/keyboard/keyboard.c 	修改：     kernel/drivers/lpt/lpt.c 	修改：     kernel/drivers/mouse/mouse.c 	修改：     kernel/drivers/pci/pci.c 	修改：     kernel/drivers/timer/timer.c 	修改：     kernel/exec/exec.c 	修改：     kernel/fault/fault.c 	修改：     kernel/fs/fat.c 	修改：     kernel/fs/fat.obj 	修改：     kernel/fs/fatfs/diskio.c 	修改：     kernel/fs/fatfs/diskio.obj 	修改：     kernel/fs/fatfs/ff.obj 	修改：     kernel/fs/fatfs/ffconf.h 	修改：     kernel/fs/iso9660.c 	修改：     kernel/fs/vfile.c 	修改：     kernel/fs/vfs.c 	修改：     kernel/gdtidt/gdtidt.c 	修改：     kernel/gui/desktop/desktop.c 	修改：     kernel/gui/graphic/graphic.c 	修改：     kernel/gui/sheet/sheet.c 	修改：     kernel/gui/theme/theme.c 	修改：     kernel/gui/window/window.c 	修改：     kernel/include/ELF.h 	修改：     kernel/include/acpi.h 	新文件：   kernel/include/audio.h 	修改：     kernel/include/binfo.h 	删除：     kernel/include/bmp.h 	修改：     kernel/include/com.h 	修改：     kernel/include/console.h 	修改：     kernel/include/cpu.h 	删除：     kernel/include/disk.h 	修改：     kernel/include/fault.h 	删除：     kernel/include/fcntl.h 	修改：     kernel/include/fpu.h 	修改：     kernel/include/graphic.h 	删除：     kernel/include/hd.h 	修改：     kernel/include/image.h 	修改：     kernel/include/ime.h 	修改：     kernel/include/int.h 	修改：     kernel/include/io.h 	删除：     kernel/include/jpeg.h 	修改：     kernel/include/keyboard.h 	修改：     kernel/include/krnlcons.h 	修改：     kernel/include/limits.h 	修改：     kernel/include/list.h 	修改：     kernel/include/lpt.h 	修改：     kernel/include/math.h 	修改：     kernel/include/message.h 	修改：     kernel/include/msdemo.h 	修改：     kernel/include/pci.h 	修改：     kernel/include/power.h 	新文件：   kernel/include/sb16.h 	新文件：   kernel/include/screenshot.h 	修改：     kernel/include/sheet.h 	修改：     kernel/include/stb_image.h 	新文件：   kernel/include/stb_image_write.h 	新文件：   kernel/include/stb_truetype.h 	修改：     kernel/include/string.h 	修改：     kernel/include/syscall.h 	修改：     kernel/include/sysset.h 	修改：     kernel/include/task.h 	删除：     kernel/include/time.h 	修改：     kernel/include/timer.h 	修改：     kernel/include/vfile.h 	修改：     kernel/include/vfs.h 	新文件：   kernel/include/wav.h 	修改：     kernel/int/int.c 	修改：     kernel/krnlcons/krnlcons.c 	修改：     kernel/lib/data_structures/Makefile 	修改：     kernel/lib/data_structures/rbtree-strptr.c 	修改：     kernel/lib/data_structures/slist-strptr.c 	修改：     kernel/lib/math/Makefile 	修改：     kernel/lib/math/math.c 	修改：     kernel/lib/sprintf/Makefile 	修改：     kernel/lib/sprintf/sprintf.c 	修改：     kernel/lib/string/Makefile 	修改：     kernel/lib/string/string.c 	修改：     kernel/main/main.c 	修改：     kernel/mm/mm.c 	修改：     kernel/plug-in/Makefile 	删除：     kernel/plug-in/bmp.asm 	修改：     kernel/plug-in/ime.c 	修改：     kernel/plug-in/ini.c 	删除：     kernel/plug-in/jpeg.c 	修改：     kernel/plug-in/msdemo.c 	新文件：   kernel/plug-in/screenshot.c 	新文件：   kernel/plug-in/wav.c 	修改：     kernel/power/power.c 	修改：     kernel/syscall/syscall.asm 	修改：     kernel/syscall/syscall.c 	修改：     kernel/task/task.c 	修改：     kernel/user/console/console.c 	修改：     kernel/user/fullscreen/fullscreen.c 	修改：     kernel/user/message/message.c 	修改：     loader/loader.asm 	修改：     preload/Makefile 	修改：     preload/base/Makefile 	新文件：   preload/base/krnlcons.c 	修改：     preload/base/preload.c 	修改：     preload/fs/fat16.c 	修改：     preload/graphic/graphic.c 	修改：     preload/include/binfo.h 	修改：     preload/include/graphic.h 	新文件：   preload/include/krnlcons.h 	修改：     preload/lib/sprintf/Makefile 	修改：     preload/lib/sprintf/sprintf.c 	修改：     preload/lib/string/Makefile 	修改：     preload/mm/mm.c 	修改：     res/neumann.ini 	修改：     res/pymb2.dat 	新文件：   res/test.wav 	修改：     res/themes/summer.jpg 	修改：     res/themes/summer.png 	删除：     symbol.txt 	删除：     symbols.txt 	修改：     tools/ftcopy.c

+++ kernel/include/timer.h
@@ -8,7 +8,7 @@ Copyright W24 Studio
 #define TIMER_H
 #include <stdint.h>
 void init_timer(uint32_t freq);
-void sleep(uint32_t timer);
-void clock_sleep(uint32_t timer);
+void sleep(uint64_t timer);
+void clock_sleep(uint64_t timer);
 int benchcpu(void);
 #endif
